+++ ecs-cli/modules/cli/local/options/options.go
@@ -17,9 +17,7 @@ package options
 import (
 	"fmt"
 
-	"github.com/aws/amazon-ecs-cli/ecs-cli/modules/cli/local/localproject"
 	"github.com/aws/amazon-ecs-cli/ecs-cli/modules/commands/flags"
-	"github.com/sirupsen/logrus"
 	"github.com/urfave/cli"
 )
 
@@ -55,22 +53,4 @@ func ValidateFlagPairs(c *cli.Context) error {
 		}
 	}
 	return nil
-}
-
-// ContainerSearchInfo show the task definition filepath or arn/family message on local down.
-func ContainerSearchInfo(c *cli.Context) {
-	if c.IsSet(flags.TaskDefinitionFile) {
-		logrus.Infof("Searching for containers matching --%s=%s", flags.TaskDefinitionFile, c.String(flags.TaskDefinitionFile))
-		return
-	}
-	if c.IsSet(flags.TaskDefinitionRemote) {
-		logrus.Infof("Searching for containers matching --%s=%s", flags.TaskDefinitionRemote, c.String(flags.TaskDefinitionRemote))
-		return
-	}
-	if c.Bool(flags.All) {
-		logrus.Info("Searching for all running containers")
-		return
-	}
-	logrus.Infof("Searching for containers matching --%s=%s", flags.TaskDefinitionFile, localproject.LocalInFileName)
-	return
 }
